changed into functional component
