+++ .expo/packager-info.json
@@ -1,10 +1,10 @@
 {
   "expoServerPort": 19000,
-  "expoServerNgrokUrl": null,
-  "packagerNgrokUrl": null,
-  "ngrokPid": null,
+  "expoServerNgrokUrl": "https://pp-g98.anonymous.ovuli.exp.direct",
+  "packagerNgrokUrl": "https://packager.pp-g98.anonymous.ovuli.exp.direct",
+  "ngrokPid": 7664,
   "devToolsPort": 19002,
   "packagerPort": 19001,
-  "packagerPid": 13692,
+  "packagerPid": 12608,
   "webpackServerPort": null
 }
