--- src/screens/HomeScreen/Calendar/CalendarHeaderComponent.js
@@ -1,143 +1,0 @@
-import React from 'react';
-import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
-import PropTypes from 'prop-types';
-import moment from 'moment';
-import { MaterialIcons } from '@expo/vector-icons';
-import { LinearGradient } from 'expo-linear-gradient';
-
-const weekDaysNames = ['S', 'M', 'T', 'W', 'T', 'F', 'S'];
-const now = moment();
-const styles = StyleSheet.create({
-  header: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    padding: 12,
-    backgroundColor: 'transparent',
-    marginLeft: 15, //Dimensions.get("screen").width - 2,
-    marginRight: 15,
-  },
-  linearGradient: {
-    flex: 1,
-    justifyContent: 'space-around',
-    flexDirection: 'row',
-  },
-  week: {
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    alignItems: 'center',
-    height: 40,
-  },
-  weekDaysNamesStyle: {
-    flex: 1,
-    flexDirection: 'row',
-    marginLeft: 7,
-    marginRight: 7,
-    justifyContent: 'space-around',
-  },
-  weekName: {
-    marginTop: 7,
-    marginBottom: 7,
-    width: 40,
-    textAlign: 'center',
-    fontSize: 20,
-    color: '#ffffff',
-  },
-
-  monthText: {
-    fontWeight: 'bold',
-    fontSize: 20,
-  },
-  yearText: {
-    fontSize: 16,
-  },
-  iconContainer: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginLeft: 8,
-    marginRight: 4,
-    marginTop: -2,
-  },
-  leftIcon: {
-    transform: [{ rotate: '180deg' }],
-  },
-  icon: {
-    width: 24,
-    height: 24,
-  },
-  disabled: {
-    opacity: 0.4,
-  },
-});
-class CalendarHeaderComponent extends React.PureComponent {
-  constructor(props) {
-    super(props);
-    this.onPressArrowLeft = this.onPressArrowLeft.bind(this);
-    this.onPressArrowRight = this.onPressArrowRight.bind(this);
-    this.shouldLeftArrowBeDisabled = this.shouldLeftArrowBeDisabled.bind(this);
-  }
-  onPressArrowLeft() {
-    this.props.onPressArrowLeft(this.props.month, this.props.addMonth);
-  }
-
-  onPressArrowRight() {
-    this.props.onPressArrowRight(this.props.month, this.props.addMonth);
-  }
-
-  shouldLeftArrowBeDisabled() {
-    const selectedDate = moment(this.props.month.getTime());
-    return selectedDate.isSame(now, 'month');
-  }
-
-  render() {
-    const month = moment(this.props.month.getTime())
-      .format('MMMM')
-      .toString();
-    const year = moment(this.props.month.getTime()).format('YYYY');
-    return (
-      <View>
-        <View style={styles.header}>
-          <TouchableOpacity onPress={this.onPressArrowLeft}>
-            <MaterialIcons name="keyboard-arrow-left" size={32} color="grey" />
-          </TouchableOpacity>
-          <Text style={styles.monthText}>
-            {month} &nbsp;
-            <Text style={styles.yearText}>{year}</Text>
-          </Text>
-
-          <TouchableOpacity style={styles.iconContainer} onPress={this.onPressArrowRight}>
-            <MaterialIcons name="keyboard-arrow-right" size={32} color="grey" />
-          </TouchableOpacity>
-        </View>
-        <View style={styles.week}>
-          <LinearGradient
-            start={{ x: 0, y: 0 }}
-            end={{ x: 1, y: 0 }}
-            colors={['#f54c63', '#f65e63', '#f78161']}
-            style={styles.linearGradient}
-          >
-            <View style={styles.weekDaysNamesStyle}>
-              {weekDaysNames.map((day, index) => (
-                <Text key={index} style={styles.weekName} numberOfLines={1}>
-                  {day}
-                </Text>
-              ))}
-            </View>
-          </LinearGradient>
-        </View>
-      </View>
-    );
-  }
-}
-
-CalendarHeaderComponent.propTypes = {
-  headerData: PropTypes.object,
-  horizontal: PropTypes.bool,
-  onPressArrowRight: PropTypes.func.isRequired,
-  onPressArrowLeft: PropTypes.func.isRequired,
-  onPressListView: PropTypes.func.isRequired,
-  onPressGridView: PropTypes.func.isRequired,
-  addMonth: PropTypes.func,
-  month: PropTypes.object,
-};
-
-export default CalendarHeaderComponent;
